Fixed field cvs
Reworked field cvs as they were not working.  Turns out there were problems with
the interpolator as well as problems with the derivative wrt to the low dimensional
coordinates of the fields.  These have now been fixed.  In addition, I have added an
action that allows you to use the interpolation objects to take in a function from a
grid that is calculated by some function and to interpolate said function onto a finer
grained grid.