Improved splines
The new spline algorithm is linear in the coefficients and allow
no errors for low order polynomial functions. Whereas this should
be irrelevant for fine grid, sum-of-Gaussian potentials, it
improves the accuracy of potentials defined on a coarse grid.

Notice that this breaks a few regtests. The fix of those is in a
separate commit.

Also notice that it is possible to revert back to old splines
by setting the environment variabled PLUMED_OLDSPLINES

Set old behavior with:
export PLUMED_OLDSPLINES=1

Go back to new behavior with:
export -n PLUMED_OLDSPLINES